+++ CHANGELOG.md
@@ -45,6 +45,7 @@ v1.8.0 [unreleased]
 - [#317](https://github.com/latera/homs/pull/317) Validate `hbw.yml` config.
 - [#318](https://github.com/latera/homs/pull/318) Improve scroll style in select_table.
 - [#327](https://github.com/latera/homs/pull/327) Optimize Camunda variables fetching.
+- [#330](https://github.com/latera/homs/pull/330) Add new available tasks components to mount.
 
 v1.7.7 [unreleased]
 -------------------